Add failing test to demonstrate unicode mishandling
As mentioned in https://github.com/richardc/perl-pod-coverage/issues/2
Pod::Parser doesn't handle encodings, which may cause some form of
failure.

Here in order to understand this, and in the attempt to add a failing
test that can be fixed with the correct behaviour, we add a simple
test fixture UnicodeHeading that refers to U+1F3A9, the Top Hat in a pod
heading.

At present it doesn't fail, so more work needed.